jobber-be/requirements.txt
duckduckgo-search
langchain-community
cohere
python-dotenv
